remove TodoList and lower todo form logic to Todos
